Email customer on CheckOut added

+++ OnlineShopping.Domain/Entities/OnlineTransactionDetail.cs
@@ -15,6 +15,6 @@ namespace OnlineShopping.Domain.Entities
         public string barcode { get; set; }
         public int shopID { get; set; }
         public int unitSold { get; set; }
-        public float totalCost { get; set; }
+        public decimal totalCost { get; set; }
     }
 }
