+++ OnlineShopping.Domain/Entities/Product.cs
@@ -6,7 +6,6 @@ using System.Threading.Tasks;
 using System.ComponentModel.DataAnnotations;
 using System.Web.Mvc;
 
-
 namespace OnlineShopping.Domain.Entities
 {
     public class Product
@@ -17,11 +16,11 @@ namespace OnlineShopping.Domain.Entities
         public string barcode { get; set; }
         public int categoryID { get; set; }
         public int manufacturerID { get; set; }
-        public float costPrice { get; set; }
-        public float maxPrice { get; set; }
+        public decimal costPrice { get; set; }
+        public decimal maxPrice { get; set; }
         public int currentStock { get; set; }
         public int minimumStock { get; set; }
         public int bundleUnit { get; set; }
-        public float discountPercentage { get; set; }
+        public decimal discountPercentage { get; set; }
     }
 }
